Completes max correlation analysis with some reformatting
The last updated analysis seemed to depend on the order we ran code in
due to the definition of Brani's code that we are using. As a result I
moved that function definition into a "preamble" of the script. That
function also appears to depend on the `wavelet` package, so I added a
`require(wavelet)` line to the start of the script. As a last bit of
formatting I have stripped away some of the whitespace in the file. This
is my taste only, but it makes jumping around in the file a lot easier.

The substantive changes are to calculate correlations based on WeiWei's
results. These deviate slightly from the case study we are following in
names, but the case study does not appear to use those names extensively
later in code. Compare the case study code with the list object defined
in Step 7 of our analysis to continue.